requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
python-multipart
jinja2